requirements.txt
numpy
fastapi
uvicorn
#tensorflow
tensorflow-cpu
tensorflow_hub
pydantic
#opencv-python
opencv-python-headless
python-multipart
gunicorn==19.9.0
uvloop
httptools